override the CSS style for the property decorator
Originally, "display: inline-block" is used in the CSS style of ".rst-content dl:not(.docutils) .property", which is defined in theme.css. This caused rendering issues such that the documentation for a method does not stretch the entire page width. As such, the documentation for class methods with the property decorator could appear next to other methods instead of underneath.