fix "SyntaxWarning: invalid escape sequence" in docstrings
Using invalid escape sequences in string literals has been deprecated since Python 3.6. A DeprecationWarning was used to issue the warning. These warnings are silent by default.

Python 3.12 changed the DeprecationWarning to a SyntaxWarning. SyntaxWarnings are emitted by the compiler when the code is parsed.

In a future Python version, SyntaxError will eventually be raised, instead of SyntaxWarning.